Move exception handling for delete()
