+++ tests/Functional/Services/MachineProvider/DigitalOceanMachineProviderTest.php
@@ -5,23 +5,14 @@ declare(strict_types=1);
 namespace App\Tests\Functional\Services\MachineProvider;
 
 use App\Entity\Worker;
-use App\Exception\MachineProvider\DigitalOcean\ApiLimitExceededException;
-use App\Exception\MachineProvider\DigitalOcean\HttpException;
-use App\Exception\MachineProvider\Exception;
-use App\Exception\MachineProvider\ExceptionInterface;
-use App\Model\MachineProviderActionInterface;
 use App\Model\ProviderInterface;
 use App\Services\MachineProvider\DigitalOceanMachineProvider;
 use App\Services\WorkerFactory;
 use App\Tests\AbstractBaseFunctionalTest;
 use App\Tests\Services\HttpResponseFactory;
 use DigitalOceanV2\Entity\Droplet as DropletEntity;
-use DigitalOceanV2\Exception\ApiLimitExceededException as VendorApiLimitExceededExceptionAlias;
-use DigitalOceanV2\Exception\RuntimeException;
-use DigitalOceanV2\Exception\ValidationFailedException;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
 use webignition\ObjectReflector\ObjectReflector;
 
 class DigitalOceanMachineProviderTest extends AbstractBaseFunctionalTest
@@ -121,83 +112,5 @@ class DigitalOceanMachineProviderTest extends AbstractBaseFunctionalTest
         $this->machineProvider->remove($this->worker);
 
         self::expectNotToPerformAssertions();
-    }
-
-    /**
-     * @dataProvider apiActionThrowsExceptionDataProvider
-     *
-     * @param class-string $expectedExceptionClass
-     */
-    public function testRemoveThrowsException(
-        ResponseInterface $apiResponse,
-        string $expectedExceptionClass,
-        \Exception $expectedRemoveException
-    ): void {
-        $this->doActionThrowsExceptionTest(
-            function () {
-                $this->machineProvider->remove($this->worker);
-            },
-            MachineProviderActionInterface::ACTION_DELETE,
-            $apiResponse,
-            $expectedExceptionClass,
-            $expectedRemoveException
-        );
-    }
-
-    /**
-     * @param MachineProviderActionInterface::ACTION_* $action
-     * @param class-string $expectedExceptionClass
-     */
-    private function doActionThrowsExceptionTest(
-        callable $callable,
-        string $action,
-        ResponseInterface $apiResponse,
-        string $expectedExceptionClass,
-        \Throwable $expectedRemoteException
-    ): void {
-        $this->mockHandler->append($apiResponse);
-
-        try {
-            $callable();
-            self::fail(ExceptionInterface::class . ' not thrown');
-        } catch (Exception $exception) {
-            self::assertSame($expectedExceptionClass, $exception::class);
-            self::assertSame($action, $exception->getAction());
-            self::assertEquals($expectedRemoteException, $exception->getRemoteException());
-        }
-    }
-
-    /**
-     * @return array[]
-     */
-    public function apiActionThrowsExceptionDataProvider(): array
-    {
-        return [
-            VendorApiLimitExceededExceptionAlias::class => [
-                'apiResponse' => new Response(
-                    429,
-                    [
-                        'RateLimit-Reset' => 123,
-                    ]
-                ),
-                'expectedExceptionClass' => ApiLimitExceededException::class,
-                'expectedRemoteException' => new VendorApiLimitExceededExceptionAlias('Too Many Requests', 429),
-            ],
-            RuntimeException::class . ' HTTP 503' => [
-                'apiResponse' => new Response(503),
-                'expectedExceptionClass' => HttpException::class,
-                'expectedRemoteException' => new RuntimeException('Service Unavailable', 503),
-            ],
-            ValidationFailedException::class => [
-                'apiResponse' => new Response(400),
-                'expectedExceptionClass' => Exception::class,
-                'expectedRemoteException' => new ValidationFailedException('Bad Request', 400),
-            ],
-            'droplet does not exist' => [
-                'apiResponse' => new Response(404),
-                'expectedExceptionClass' => HttpException::class,
-                'expectedRemoteException' => new RuntimeException('Not Found', 404),
-            ],
-        ];
     }
 }
